test: merge executor's serial tests to other test
