+++ executor/slow_query_test.go
@@ -452,7 +452,7 @@ select 7;`
 	sctx.GetSessionVars().TimeZone = loc
 	sctx.GetSessionVars().SlowQueryFile = fileName3
 	for i, cas := range cases {
-		extractor := &plannercore.SlowQueryExtractor{Enable: (len(cas.startTime) > 0 && len(cas.endTime) > 0)}
+		extractor := &plannercore.SlowQueryExtractor{Enable: len(cas.startTime) > 0 && len(cas.endTime) > 0}
 		if extractor.Enable {
 			startTime, err := ParseTime(cas.startTime)
 			c.Assert(err, IsNil)
@@ -622,7 +622,7 @@ select 9;`
 	sctx.GetSessionVars().TimeZone = loc
 	sctx.GetSessionVars().SlowQueryFile = fileName3
 	for i, cas := range cases {
-		extractor := &plannercore.SlowQueryExtractor{Enable: (len(cas.startTime) > 0 && len(cas.endTime) > 0), Desc: true}
+		extractor := &plannercore.SlowQueryExtractor{Enable: len(cas.startTime) > 0 && len(cas.endTime) > 0, Desc: true}
 		if extractor.Enable {
 			startTime, err := ParseTime(cas.startTime)
 			c.Assert(err, IsNil)
